requirements.txt
asgiref==3.2.10
atomicwrites==1.4.0
attrs==20.2.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
flake8==3.8.3
idna==2.10
iniconfig==1.0.1
Markdown==3.2.2
mccabe==0.6.1
more-itertools==8.5.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pyflakes==2.2.0
pyparsing==2.4.7
pytest==6.0.2
pytest-django==3.10.0
python-decouple==3.3
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
